requirements.txt
flask==2.3.3
py-algorand-sdk==2.6.0
pyteal==0.25.0
python-dotenv==1.0.0
